+++ carousel-kit.php
@@ -10,7 +10,7 @@
  * Contributors: iamdanih17, immasud
  * License:     GPL2
  * License URI: https://www.gnu.org/licenses/gpl-2.0.html
- * Version:     1.0.0
+ * Version:     1.0.3
  * Text Domain: carousel-kit
  *
  * @package carousel-kit